feat: v[i] on a packed array through a typed view (#2041); Format-tier stdlib docs (#1523)
#2041 was recorded as blocked on a distinct handle type threaded through
three module APIs and every caller. It is not. `[]` genuinely cannot
dispatch on the HANDLE -- a bare ptr has no element type -- but the handle
is not the only thing you can index, and std.strarr has shipped the answer
since it was written. intarr.array(a) returns an int[] over the same
buffer, so v[3] = 42 works, and after #1986's inlining it lowers to the
same load the accessor does. The [] diagnostic now names the view first,
and the asks/ reply that deferred the ask carries an update saying its
conclusion was wrong.

docs/stdlib-reference.md gains the six Format-tier modules -- bits, hash,
bignum, msgpack, cbor, yaml -- taking the doc gate from 73 run-and-match
blocks to 78. yaml's is compile-checked rather than run: it needs libfyaml
at build time, so its output depends on how the toolchain was built.

Two defects found on the way. std.longarr said `long long*` where Aether's
`long` lowers to int64_t -- on LP64 those are different pointer types, a
hard error on gcc 14, invisible on Windows. And fmt_gate left its
in-tree sibling behind when killed: one such leftover had already been
committed to main in #2165 and was being run as a test.

+++ std/collections/aether_longarr.c
@@ -6,9 +6,9 @@
  * (hash state words, crypto round buffers, wide DP tables) where going
  * through std.list's void*-boxed items would cost an allocation per
  * entry and chase an extra pointer. Aether's `long` lowers to C
- * `long long`, so the element type is `long long` end-to-end.
+ * `int64_t`, so the element type is `int64_t` end-to-end.
  *
- * Shape is as minimal as it gets: heap-allocate size*long long, store a
+ * Shape is as minimal as it gets: heap-allocate size*int64_t, store a
  * size next to it, expose direct index ops plus a bulk-fill. No
  * amortised growth, no capacity-beyond-size. Callers that need resize
  * on write use std.list. */
@@ -21,7 +21,7 @@
 LongArray* longarr_new_raw(int size) {
     if (size < 0) return NULL;
     /* #463: cap-aware. The struct is sizeof(LongArray); the data
-     * array is size×sizeof(long long). Both byte counts are recoverable
+     * array is size×sizeof(int64_t). Both byte counts are recoverable
      * at free time from `arr->size`. */
     LongArray* arr = (LongArray*)aether_caps_malloc(sizeof(*arr));
     if (!arr) return NULL;
@@ -30,12 +30,12 @@ LongArray* longarr_new_raw(int size) {
         arr->data = NULL;   // legal empty array; longarr_free still OK
         return arr;
     }
-    arr->data = (long long*)aether_caps_calloc((size_t)size, sizeof(long long));
+    arr->data = (int64_t*)aether_caps_calloc((size_t)size, sizeof(int64_t));
     if (!arr->data) { aether_caps_free(arr, sizeof(*arr)); return NULL; }
     return arr;
 }
 
-LongArray* longarr_new_filled_raw(int size, long long init) {
+LongArray* longarr_new_filled_raw(int size, int64_t init) {
     LongArray* arr = longarr_new_raw(size);
     if (!arr) return NULL;
     for (int i = 0; i < size; i++) arr->data[i] = init;
@@ -46,12 +46,12 @@ int longarr_size(LongArray* arr) {
     return arr ? arr->size : -1;
 }
 
-long long longarr_get_raw(LongArray* arr, int i) {
+int64_t longarr_get_raw(LongArray* arr, int i) {
     if (!arr || i < 0 || i >= arr->size) return 0;
     return arr->data[i];
 }
 
-void longarr_set_raw(LongArray* arr, int i, long long value) {
+void longarr_set_raw(LongArray* arr, int i, int64_t value) {
     if (!arr || i < 0 || i >= arr->size) return;
     arr->data[i] = value;
 }
@@ -60,15 +60,15 @@ void longarr_set_raw(LongArray* arr, int i, long long value) {
  * aether_collections.h (#1986), for the same reason as the int and float
  * twins: the definition must be visible where the indexing happens. */
 
-void longarr_fill(LongArray* arr, long long value) {
+void longarr_fill(LongArray* arr, int64_t value) {
     if (!arr || arr->size == 0) return;
     for (int i = 0; i < arr->size; i++) arr->data[i] = value;
 }
 
 void longarr_free(LongArray* arr) {
     if (!arr) return;
-    /* data was calloc'd size×sizeof(long long) (NULL for the empty
+    /* data was calloc'd size×sizeof(int64_t) (NULL for the empty
      * array, where caps_free is a no-op). */
-    aether_caps_free(arr->data, (size_t)arr->size * sizeof(long long));
+    aether_caps_free(arr->data, (size_t)arr->size * sizeof(int64_t));
     aether_caps_free(arr, sizeof(*arr));
 }
